// react/lib/schematic/api/checkoutexternal/models/BillingProductPriceTierResponseData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schematic API
 * Schematic API
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BillingProductPriceTierResponseData
 */
export interface BillingProductPriceTierResponseData {
    /**
     * 
     * @type {number}
     * @memberof BillingProductPriceTierResponseData
     */
    flatAmount?: number | null;
    /**
     * 
     * @type {number}
     * @memberof BillingProductPriceTierResponseData
     */
    perUnitPrice?: number | null;
    /**
     * 
     * @type {string}
     * @memberof BillingProductPriceTierResponseData
     */
    perUnitPriceDecimal?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BillingProductPriceTierResponseData
     */
    upTo?: number | null;
}

/**
 * Check if a given object implements the BillingProductPriceTierResponseData interface.
 */
export function instanceOfBillingProductPriceTierResponseData(value: object): value is BillingProductPriceTierResponseData {
    return true;
}

export function BillingProductPriceTierResponseDataFromJSON(json: any): BillingProductPriceTierResponseData {
    return BillingProductPriceTierResponseDataFromJSONTyped(json, false);
}

export function BillingProductPriceTierResponseDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): BillingProductPriceTierResponseData {
    if (json == null) {
        return json;
    }
    return {
        
        'flatAmount': json['flat_amount'] == null ? undefined : json['flat_amount'],
        'perUnitPrice': json['per_unit_price'] == null ? undefined : json['per_unit_price'],
        'perUnitPriceDecimal': json['per_unit_price_decimal'] == null ? undefined : json['per_unit_price_decimal'],
        'upTo': json['up_to'] == null ? undefined : json['up_to'],
    };
}

export function BillingProductPriceTierResponseDataToJSON(json: any): BillingProductPriceTierResponseData {
    return BillingProductPriceTierResponseDataToJSONTyped(json, false);
}

export function BillingProductPriceTierResponseDataToJSONTyped(value?: BillingProductPriceTierResponseData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'flat_amount': value['flatAmount'],
        'per_unit_price': value['perUnitPrice'],
        'per_unit_price_decimal': value['perUnitPriceDecimal'],
        'up_to': value['upTo'],
    };
}