// react/lib/schematic/api/checkoutexternal/models/CompanySubscriptionResponseData.ts
export interface CompanySubscriptionResponseData {
    /**
     * 
     * @type {Date}
     * @memberof CompanySubscriptionResponseData
     */
    cancelAt?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof CompanySubscriptionResponseData
     */
    cancelAtPeriodEnd: boolean;
    /**
     * 
     * @type {string}
     * @memberof CompanySubscriptionResponseData
     */
    currency: string;
    /**
     * 
     * @type {string}
     * @memberof CompanySubscriptionResponseData
     */
    customerExternalId: string;
    /**
     * 
     * @type {Array<BillingSubscriptionDiscountView>}
     * @memberof CompanySubscriptionResponseData
     */
    discounts: Array<BillingSubscriptionDiscountView>;
    /**
     * 
     * @type {Date}
     * @memberof CompanySubscriptionResponseData
     */
    expiredAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof CompanySubscriptionResponseData
     */
    interval: string;
    /**
     * 
     * @type {boolean}
     * @memberof CompanySubscriptionResponseData
     */
    isInitial: boolean;
    /**
     * 
     * @type {InvoiceResponseData}
     * @memberof CompanySubscriptionResponseData
     */
    latestInvoice?: InvoiceResponseData;
    /**
     * 
     * @type {PaymentMethodResponseData}
     * @memberof CompanySubscriptionResponseData
     */
    paymentMethod?: PaymentMethodResponseData;
    /**
     * 
     * @type {Array<BillingProductForSubscriptionResponseData>}
     * @memberof CompanySubscriptionResponseData
     */
    products: Array<BillingProductForSubscriptionResponseData>;
    /**
     * 
     * @type {BillingProviderType}
     * @memberof CompanySubscriptionResponseData
     */
    providerType: BillingProviderType;
    /**
     * 
     * @type {string}
     * @memberof CompanySubscriptionResponseData
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof CompanySubscriptionResponseData
     */
    subscriptionExternalId: string;
    /**
     * 
     * @type {number}
     * @memberof CompanySubscriptionResponseData
     */
    totalPrice: number;
    /**
     * 
     * @type {Date}
     * @memberof CompanySubscriptionResponseData
     */
    trialEnd?: Date | null;
}